+++ ThomasTheTankEngine/Systems/UpdateOnly/InputSystem.hpp
@@ -14,7 +14,7 @@
 class InputSystem : public System {
 public:
     void init();
-    void tick(float dt);
+    void tick(uint64_t dt);
     void render(){};
     void teardown();
